@@ -5,13 +5,13 @@ class Product {
   final int price, size, id;
   final Color color;
   Product({
-    this.id,
-    this.image,
-    this.title,
-    this.price,
-    this.description,
-    this.size,
-    this.color,
+    required this.id,
+    required this.image,
+    required this.title,
+    required this.price,
+    required this.description,
+    required this.size,
+    required this.color,
   });
 }
 
