@@ -16,10 +16,10 @@ class Body extends StatelessWidget {
           padding: const EdgeInsets.symmetric(horizontal: kDefaultPaddin),
           child: Text(
             "Women",
-            style: Theme.of(context)
-                .textTheme
-                .headline5
-                .copyWith(fontWeight: FontWeight.bold),
+            // style: Theme.of(context)
+            //     .textTheme
+            //     .headline5
+            //     .copyWith(fontWeight: FontWeight.bold),
           ),
         ),
         Categories(),
